+++ tests/unittests/objects/failure_rule/test_firewatch_objects_failure_rule.py
@@ -1,7 +1,5 @@
 import unittest
 
-import pytest
-
 from cli.objects.failure_rule import FailureRule
