+++ tests/unittests/objects/job/test_firewatch_objects_job_find_failures.py
@@ -1,6 +1,5 @@
 import tempfile
 
-from cli.objects.configuration import Configuration
 from cli.objects.job import Job
 from tests.unittests import helpers
 from tests.unittests.objects.job.job_base_test import JobBaseTest
