+++ tests/unittests/objects/job/test_firewatch_objects_job_get_download_path.py
@@ -1,6 +1,5 @@
 from unittest.mock import patch
 
-from cli.objects.configuration import Configuration
 from cli.objects.job import Job
 from tests.unittests.objects.job.job_base_test import JobBaseTest
