+++ tests/unittests/objects/rule/test_firewatch_objects_rule_get_jira_additional_labels.py
@@ -2,7 +2,6 @@ from unittest.mock import patch
 
 import pytest
 
-from cli.objects.rule import Rule
 from tests.unittests.objects.rule.rule_base_test import RuleBaseTest
