+++ tests/unittests/objects/rule/test_firewatch_objects_rule_get_jira_affects_version.py
@@ -1,6 +1,5 @@
 from unittest.mock import patch
 
-from cli.objects.rule import Rule
 from tests.unittests.objects.rule.rule_base_test import RuleBaseTest
